+++ internal/app/common.go
@@ -1,31 +1,28 @@
 package app
 
 import (
-	"context"
+	// "context"
 
 	"github.com/gitopia/git-server/internal/db"
 	"github.com/gitopia/git-server/utils"
-	gitopiatypes "github.com/gitopia/gitopia/v4/x/gitopia/types"
+	// gitopiatypes "github.com/gitopia/gitopia/v5/x/gitopia/types"
 )
 
 func (s *Server) CacheRepository(repoId uint64) error {
-	res, err := s.QueryService.GitopiaRepositoryStorage(context.Background(), &gitopiatypes.QueryGetRepositoryStorageRequest{
-		RepositoryId: repoId,
-	})
-	if err != nil {
-		return err
-	}
+	// TODO: get latest packfile info from the chain
+	name := "pack-xyz.pack"
+	cid := "QmPK1xyz"
 
-	if !utils.IsCached(db.CacheDb, repoId, res.Storage.Latest.Id, res.Storage.Latest.Name) {
+	if !utils.IsCached(db.CacheDb, repoId, cid, name) {
 		CacheMutex.Lock()
 		defer CacheMutex.Unlock()
-		if err := utils.DownloadRepo(db.CacheDb, repoId, r.RepoPath, &s.Config); err != nil {
+		if err := utils.DownloadRepo(db.CacheDb, repoId, "repo-path", &s.Config); err != nil {
 			return err
 		}
 	} else { // Repository is already available in file system cache
 		CacheMutex.Lock()
 		// Increase cache expiry time for this repository
-		if err := utils.UpdateCacheEntry(db.CacheDb, repoId, res.Storage.Latest.Id, res.Storage.Latest.Name); err != nil {
+		if err := utils.UpdateCacheEntry(db.CacheDb, repoId, cid, name); err != nil {
 			return err
 		}
 		CacheMutex.Unlock()