+++ internal/app/handler/content.go
@@ -1,4 +1,4 @@
-package route
+package handler
 
 import (
 	"encoding/base64"